pyproject.toml
[project]
name = "intacct-syncer"
version = "0.1.0"
description = "Sync selected data from Intacct Production to Development and create Shippers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "click==8.1.7",
  "httpx==0.27.2",
  "python-dotenv==1.0.1",
  "tenacity==8.5.0",
  "pydantic==2.9.0",
  "structlog==24.1.0",
]

[project.scripts]
intacctsync = "intacctsync:cli"

[tool.uv]
package = true
dev-dependencies = [
  "ruff==0.6.9",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["intacctsync"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
